requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.3.0
pydantic-settings>=2.0.3
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
bertopic>=0.15.0
scikit-learn>=1.3.0
numpy>=1.25.2
